Safari-609.1.11.1/LayoutTests/http/tests/events/device-orientation-motion-non-secure-context-expected.txt

CONSOLE MESSAGE: line 38: Device Orientation API is not supported
CONSOLE MESSAGE: line 20: Device Motion API is not supported
Tests that trying to set an event listener for deviceorientation and deviceorientation logs an error in non-secure contexts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
